separated all *References out of specs.py
Currently, all *References class lives in specs.py, but we should move those out to it's own file for organizational purposes
