+++ metricflow/dataflow/dataflow_plan.py
@@ -36,12 +36,12 @@ from metricflow.dataset.dataset import DataSet
 from metricflow.model.objects.metric import CumulativeMetricWindow
 from metricflow.model.objects.elements.measure import AggregationType
 from metricflow.object_utils import pformat_big_objects
+from metricflow.references import TimeDimensionReference
 from metricflow.specs import (
     OrderBySpec,
     InstanceSpec,
     MetricSpec,
     LinklessIdentifierSpec,
-    TimeDimensionReference,
     TimeDimensionSpec,
     SpecWhereClauseConstraint,
 )
