+++ metricflow/model/objects/elements/dimension.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Optional
 
 from metricflow.model.objects.base import HashableBaseModel
-from metricflow.specs import DimensionReference, TimeDimensionReference
+from metricflow.references import DimensionReference, TimeDimensionReference
 from metricflow.time.time_granularity import TimeGranularity
 from metricflow.object_utils import ExtendedEnum
