+++ metricflow/model/validations/unique_valid_name.py
@@ -26,7 +26,7 @@ from metricflow.model.validations.validator_helpers import (
     validate_safely,
 )
 from metricflow.object_utils import assert_values_exhausted
-from metricflow.specs import ElementReference
+from metricflow.references import ElementReference
 from metricflow.time.time_granularity import TimeGranularity
